+++ packages/mobx-state-tree/__tests__/core/api.test.ts
@@ -75,6 +75,9 @@ const METHODS_AND_INTERNAL_TYPES = stringToArray(`
     isValidReference,
     tryReference,
     getNodeId,
+    getRunningActionContext,
+    isActionContextChildOf,
+    isActionContextThisOrChildOf,
 
     types
 `)